+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 *Notice: I am aware of ongoing issues with music cutting out or changing suddenly. This is under investigation!*
 
 ### Patch Highlights:
+- Fixed: Miniboss rooms in the Temple of Styx do not always offer upgraded rewards.
 
 Other fixes and improvements:
 - Fixed: The restocked item spawned by Hermes' "Travel Deal" disappears after returning to the Styx hub room if it is not bought before visiting another room.